+++ src/applications/mhv/medications/components/shared/FeedbackEmail.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const FeedbackEmail = () => {
+  return <a href="mailto: VAmhvfeedback@va.gov">VAmhvfeedback@va.gov</a>;
+};
+
+export default FeedbackEmail;
