+++ src/applications/mhv/medications/containers/LandingPage.jsx
@@ -1,124 +1,111 @@
-import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { getPrescriptionsList } from '../actions/prescriptions';
-import { setBreadcrumbs } from '../actions/breadcrumbs';
-import MedicationsList from '../components/MedicationsList/MedicationsList';
-import MedicationsListSort from '../components/MedicationsList/MedicationsListSort';
-import PrintHeader from './PrintHeader';
+import React from 'react';
+import FeedbackEmail from '../components/shared/FeedbackEmail';
 
 const LandingPage = () => {
-  const prescriptions = useSelector(
-    state => state.rx.prescriptions.prescriptionsList,
-  );
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    if (prescriptions) {
-      dispatch(
-        setBreadcrumbs([], {
-          url: '/my-health/medications/prescriptions/',
-          label: 'Medications',
-        }),
-      );
-    }
-  });
-
-  useEffect(
-    () => {
-      dispatch(getPrescriptionsList());
-    },
-    [dispatch],
-  );
-
   const content = () => {
-    if (prescriptions) {
-      return (
-        <div className="landing-page">
-          <PrintHeader />
-          <h1 className="page-title">Medications</h1>
-          <div className="vads-u-margin-bottom--2 no-print">
-            Review your prescription medicaitons from VA, and providers outside
-            of our network.
-          </div>
-          <div className="landing-page-content">
-            <div className="no-print">
-              <button
-                type="button"
-                className="link-button vads-u-display--block vads-u-margin-bottom--2"
-                data-testid="print-records-button"
-                onClick={() => window.print()}
-              >
-                <i
-                  aria-hidden="true"
-                  className="fas fa-print vads-u-margin-right--0p5"
-                />
-                Print medication list
-              </button>
-              <button
-                type="button"
-                className="link-button vads-u-display--block vads-u-margin-bottom--2"
-              >
-                <i
-                  aria-hidden="true"
-                  className="fas fa-download vads-u-margin-right--0p5"
-                />
-                Download list as a PDF
-              </button>
-              <button
-                type="button"
-                className="link-button vads-u-display--block vads-u-margin-bottom--2"
-              >
-                <i
-                  aria-hidden="true"
-                  className="fas fa-download vads-u-margin-right--0p5"
-                />
-                Download list as a Text file
-              </button>
-              <va-additional-info trigger="What to know about downloading records">
-                <ul>
-                  <li>
-                    <strong>If you’re on a public or shared computer,</strong>{' '}
-                    print your records instead of downloading. Downloading will
-                    save a copy of your records to the public computer.
-                  </li>
-                  <li>
-                    <strong>If you use assistive technology,</strong> a Text
-                    file (.txt) may work better for technology such as screen
-                    reader, screen enlargers, or Braille displays.
-                  </li>
-                </ul>
-              </va-additional-info>
-              <MedicationsListSort />
-              <div className="rx-page-total-info vads-u-border-bottom--2px vads-u-border-color--gray-lighter" />
-            </div>
-            <MedicationsList rxList={prescriptions} />
-          </div>
-          <div className="rx-landing-page-footer no-print">
-            <div className="footer-header vads-u-font-size--h2 vads-u-font-weight--bold vads-u-padding-y--1 vads-u-border-bottom--1px vads-u-border-color--gray-light">
-              Resources related to medications
-            </div>
-            <div className="footer-links">
-              <a href="nolink">Allergies and Adverse Reactions</a>
-              <p>
-                This is a description of why the user may need to navigate to
-                medical records to see their allergies.
-              </p>
-              <a href="nolink">Resources and Support</a>
-              <p>
-                This is a description of what the user might find in resources
-                and support.
-              </p>
-            </div>
-          </div>
-        </div>
-      );
-    }
     return (
-      <va-loading-indicator
-        message="Loading..."
-        setFocus
-        data-testid="loading-indicator"
-      />
+      <div className="landing-page">
+        <div className="main-content">
+          <section>
+            <h1>About Medications</h1>
+            <p className="vads-u-font-size--h3">
+              Learn how to manage your VA prescriptions and review all
+              medications in your VA medical records.
+            </p>
+          </section>
+          <section>
+            <a
+              className="vads-c-action-link--green"
+              href="/my-health/medications/prescriptions"
+            >
+              Go to your medications
+            </a>
+          </section>
+          <section>
+            <h2>What to know as you try out this tool</h2>
+            <p>
+              We’re giving the trusted My HealtheVet pharmacy tool a new home
+              here on VA.gov. And we need your feedback to help us keep making
+              this tool better for you and all Veterans.
+            </p>
+            <p>
+              Email your feedback and questions to us at <FeedbackEmail />.
+            </p>
+            <p>
+              Note: You still have access to the pharmacy tool on the My
+              HealtheVet website. You can go back to that site at any time.{' '}
+              <a
+                href="/my-health/medications/"
+                target="_blank"
+                rel="noreferrer"
+              >
+                Go back to pharmacy on the My HealtheVet website
+              </a>
+            </p>
+          </section>
+          <section>
+            <h2>Questions about your medical records</h2>
+            <va-accordion bordered>
+              <va-accordion-item>
+                <h3 className="vads-u-font-size--h6" slot="headline">
+                  Does this tool list all my medications and supplies?
+                </h3>
+              </va-accordion-item>
+              <va-accordion-item>
+                <h3 className="vads-u-font-size--h6" slot="headline">
+                  What types of prescriptions can I refill and track in this
+                  tool?
+                </h3>
+              </va-accordion-item>
+              <va-accordion-item>
+                <h3 className="vads-u-font-size--h6" slot="headline">
+                  When will I get my prescriptions, and when should I request a
+                  refill?
+                </h3>
+              </va-accordion-item>
+              <va-accordion-item>
+                <h3 className="vads-u-font-size--h6" slot="headline">
+                  Will VA protect my personal health information?
+                </h3>
+              </va-accordion-item>
+              <va-accordion-item>
+                <h3 className="vads-u-font-size--h6" slot="headline">
+                  What if I have more questions?
+                </h3>
+              </va-accordion-item>
+            </va-accordion>
+          </section>
+          <section>
+            <h2>More ways to manage your medications</h2>
+            <p>
+              {' '}
+              Learn how to renew prescriptions, update your mailing address, and
+              review allergies and reactions in your VA medical records.
+            </p>
+            <va-accordion bordered>
+              <va-accordion-item>
+                <h3 className="vads-u-font-size--h6" slot="headline">
+                  How to renew prescriptions
+                </h3>
+              </va-accordion-item>
+              <va-accordion-item>
+                <h3 className="vads-u-font-size--h6" slot="headline">
+                  How to confirm or update your mailing address
+                </h3>
+              </va-accordion-item>
+              <va-accordion-item>
+                <h3 className="vads-u-font-size--h6" slot="headline">
+                  How to review your allergies and reactions
+                </h3>
+              </va-accordion-item>
+            </va-accordion>
+          </section>
+        </div>
+        <section>
+          <hr />
+          <button className="feedback"> feedback </button>
+        </section>
+      </div>
     );
   };
