+++ src/applications/mhv/medications/routes.jsx
@@ -3,12 +3,16 @@ import { Switch, Route } from 'react-router-dom';
 import App from './containers/App';
 import PrescriptionDetails from './containers/PrescriptionDetails';
 import RxBreadcrumbs from './containers/RxBreadcrumbs';
+import LandingPage from './containers/LandingPage';
 
 const routes = (
   <div className="vads-l-grid-container main-content">
     <RxBreadcrumbs />
     <div className="main-content">
       <Switch>
+        <Route exact path="/" key="LandingPage">
+          <LandingPage />
+        </Route>
         <Route exact path="/prescriptions/" key="App">
           <App />
         </Route>
